+++ test/unit/ruby_aem_spec.rb
@@ -21,6 +21,7 @@ describe 'RubyAem' do
       expect(aem).to respond_to(:package)
       expect(aem).to respond_to(:path)
       expect(aem).to respond_to(:replication_agent)
+      expect(aem).to respond_to(:outbox_replication_agent)
       expect(aem).to respond_to(:reverse_replication_agent)
       expect(aem).to respond_to(:repository)
       expect(aem).to respond_to(:user)
@@ -105,6 +106,15 @@ describe 'RubyAem' do
     it 'should return replication agent instance' do
       replication_agent = RubyAem::Aem.new().replication_agent('author', 'some-replication-agent')
       expect(replication_agent).to_not be(nil)
+    end
+
+  end
+
+  describe 'test outbox replication agent' do
+
+    it 'should return outbox replication agent instance' do
+      outbox_replication_agent = RubyAem::Aem.new().outbox_replication_agent('publish', 'some-outbox-replication-agent')
+      expect(outbox_replication_agent).to_not be(nil)
     end
 
   end